refactor: Application 하위 패키지 구조 수정 - Handler: API와 맵핑되어 요청을 처리하는 핸들러를 관리한다. - Domain: 활용되는 도메인 객체를 관리한다. - DAO: 데이터 저장소와 상호작용하는 객체를 관리한다.
