+++ src/main/java/codesquad/application/domain/Article.java
@@ -1,4 +1,4 @@
-package codesquad.application;
+package codesquad.application.domain;
 
 public record Article(
     String title,
