+++ src/main/java/codesquad/application/domain/User.java
@@ -1,4 +1,4 @@
-package codesquad.application;
+package codesquad.application.domain;
 
 import java.util.Objects;
