+++ src/main/java/codesquad/application/handler/HtmlPageHandler.java
@@ -1,7 +1,9 @@
-package codesquad.application;
+package codesquad.application.handler;
 
 import static codesquad.webserver.file.FileHttpResponseCreator.create;
 
+import codesquad.application.domain.Article;
+import codesquad.application.domain.User;
 import codesquad.database.ArticleDatabase;
 import codesquad.webserver.annotation.Controller;
 import codesquad.webserver.annotation.RequestMapping;
