+++ src/main/java/codesquad/database/ArticleDatabase.java
@@ -1,6 +1,6 @@
 package codesquad.database;
 
-import codesquad.application.Article;
+import codesquad.application.domain.Article;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -8,6 +8,8 @@ import java.util.concurrent.CopyOnWriteArrayList;
 
 public class ArticleDatabase {
     private static final List<Article> articles = new CopyOnWriteArrayList<>();
+
+    private ArticleDatabase() {}
 
     public static void add(Article article) {
         articles.add(article);
